add step quiz review delegate

+++ app/src/main/java/org/stepik/android/domain/review_instruction/model/ReviewInstruction.kt
@@ -14,7 +14,7 @@ data class ReviewInstruction(
     @SerializedName("step")
     val step: Long,
     @SerializedName("min_reviews")
-    val minReviews: Long,
+    val minReviews: Int,
     @SerializedName("strategy_type")
     val strategyType: ReviewStrategyType,
     @SerializedName("rubrics")
